requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
dispatch==1.1.0
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
docopt==0.6.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
jsonfield==3.1.0
multidispatch==0.2
multipledispatch==0.6.0
numpy==1.19.0
openpyxl==3.0.4
pandas==1.0.5
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
XlsxWriter==1.2.9
